[TASK] Move shared functions to trait
