+++ packages/guides/src/Compiler/NodeTransformers/MenuNodeTransformers/SubInternalMenuEntryNodeTransformer.php
@@ -18,6 +18,9 @@ use function assert;
 
 class SubInternalMenuEntryNodeTransformer extends AbstractMenuEntryNodeTransformer
 {
+    use MenuEntryManagement;
+    use SubSectionHierarchyHandler;
+
     // Setting a default level prevents PHP errors in case of circular references
     private const DEFAULT_MAX_LEVELS = 10;
